requirements.txt
discord.py
PyYAML
mysql-connector-python==8.0.12